app/backend/requirements.txt
# Web Framework
Flask>=3.1.0
flask-cors>=5.0.1
uvicorn>=0.34.0

# Vector Store and Database
langchain-postgres>=0.0.13
pgvector>=0.3.6
psycopg>=3.2.6
psycopg-binary>=3.2.6

# ML and Embeddings
langchain-huggingface>=0.1.2
sentence-transformers>=4.0.1
torch>=2.6.0

# Document Processing
PyMuPDF>=1.25.4
pymupdf4llm>=0.0.19
python-pptx>=1.0.1
pandas>=2.2.3
openpyxl>=3.1.5
tiktoken>=0.9.0

# Agno Framework
agno>=0.1.0
openai>=1.1.0

# Utils
python-dotenv>=1.1.0
pydantic>=2.11.0
